Refactor the email controllers file
